Implement full REST API for event management system
Adds controllers, services, DTOs, and implementations for auditoría, autenticación, eventos, modalidades, tipos de evento, roles, participación, check-in, and certificación modules. Introduces API documentation and updates entity, DTO, and service interfaces for full CRUD and business logic, including validation and error handling. Removes obsolete files and aligns security and configuration for CORS and JWT. This establishes a complete backend structure for the E-Pa-Ce system.

+++ backend/src/main/java/com/example/backend/rol/entity/RolEntity.java
@@ -20,7 +20,7 @@ public class RolEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID_Rol", nullable = false)
-    private Short idRol;
+    private Integer idRol;
 
     @NotBlank(message = "El nombre del rol es obligatorio")
     @Size(max = 20, message = "El nombre del rol no puede exceder 20 caracteres")
@@ -30,9 +30,8 @@ public class RolEntity {
     public RolEntity() {
     }
 
-    public RolEntity(Short idRol, String nombreRol) {
+    public RolEntity(Integer idRol, String nombreRol) {
         this.idRol = idRol;
         this.nombreRol = nombreRol;
     }
 }
-
